# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Gestão Escolar Api",
        "description": "A gestão escolar com API envolve a integração de sistemas e aplicativos para otimizar a administração escolar, permitindo a troca de dados e informações de forma eficiente entre diferentes componentes do ambiente educacional, como registros acadêmicos, matrículas, sistemas de pagamento e comunicação com pais e alunos.",
        "contact": {
            "email": "adilson2012jose@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.4.0"
    },
    "paths": {
...
        "/api/Estudante/MultasePrazos/anolectivo/{anolectivoid}": {
            "get": {
                "tags": [
                    "Estudante"
                ],
                "summary": "Todos os Detalhes de Pagamento Estaram neste Request",
                "operationId": "34f7321ce356cc1b98f99be3ae10f427",
                "parameters": [
                    {
                        "name": "ano",
                        "in": "path",
                        "description": "ID do ano letivo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes Concernentes As Multas Dividas ,Meses Pago ,Meses Com Divida ,Meses Com Multa Tudo esta aqui neste Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "Professor Professor": {
                                                "type": "string",
                                                "example": "Joao Miguel"
                                            },
                                            "nomeCurso": {
                                                "type": "string",
                                                "example": "Ciências Físicas e Biológica"
                                            },
                                            "nomeDisciplina": {
                                                "type": "string",
                                                "example": "INGLÊS"
                                            },
                                            "nomePeriodo": {
                                                "type": "string",
                                                "example": "Tarde"
                                            },
                                            "nomeSala": {
                                                "type": "string",
                                                "example": "Sala 9"
                                            },
                                            "nomeTurma": {
                                                "type": "string",
                                                "example": "B"
                                            },
                                            "classe_name": {
                                                "type": "string",
                                                "example": "10 Classe"
                                            },
                                            "ano_lectivo": {
                                                "type": "string",
                                                "example": "2023\\/2024"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/Encarregado/encarregadoconsultarfaltas/anolectivo/{anolectivo}/estudante/{id}": {
            "get": {
                "tags": [
                    "Encarregado"
                ],
                "summary": "lista de Faltas de um estudante durante o ano lectivo em uma certa disciplina",
                "operationId": "ba5d4b76b15f8474484b9986f7b8e1b5",
                "parameters": [
                    {
                        "name": "ano",
                        "in": "path",
                        "description": "ID do ano letivo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "estudanteid",
                        "in": "path",
                        "description": "ID do estudante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "lista de faltas por disciplinas durante o ano lectivo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "disciplina_id": {
                                                "type": "integer",
                                                "example": 15
                                            },
                                            "attendance_Nome": {
                                                "type": "string",
                                                "example": "Ausente"
                                            },
                                            "nomeDisciplina": {
                                                "type": "string",
                                                "example": "EDUCAÇÃO FISICA"
                                            },
                                            "classe_name": {
                                                "type": "string",
                                                "example": "10 Classe"
                                            },
                                            "ano_lectivo": {
                                                "type": "string",
                                                "example": "2023\\/2024"
                                            },
                                            "ausente_count": {
                                                "type": "integer",
                                                "example": "3"
                                            },
                                            "presente_count": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "atrasado_count": {
                                                "type": "integer",
                                                "example": "0"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}